+++ packages/core/src/languages/html/css-classes.ts
@@ -1,19 +1,25 @@
 import type { ManglerExpression } from "../../types";
 
-import { SingleGroupManglerExpression } from "../utils/mangler-expressions";
+import { NestedGroupExpression } from "../utils/mangler-expressions";
 
 const GROUP_NAME_MAIN = "main";
 
 const pattern: ManglerExpression[] = [
   // Finds e.g., "cls-a" and "cls-b" in  `<div class="cls-a ignore cls-b">`
-  ...["\"", "'"].map((quote) => new SingleGroupManglerExpression(
+  ...["\"", "'"].map((quote) => new NestedGroupExpression(
     `
-      (?<=
-        class\\s*=\\s*${quote}
-        ([^${quote}]*\\s)?
+      (?<=class\\s*=\\s*${quote})
+      (?<${GROUP_NAME_MAIN}>
+        (?:[^${quote}]+\\s)?
+        %s
+        (?:\\s[^${quote}]+)?
       )
+      (?=${quote})
+    `,
+    `
+      (?<=\\s|^)
       (?<${GROUP_NAME_MAIN}>%s)
-      (?=\\s|${quote})
+      (?=\\s|$)
     `,
     GROUP_NAME_MAIN,
   )),